+++ src/dataPowers/Coin.java
@@ -15,7 +15,7 @@ public class Coin extends PowerUp
 	// determina se l'oggetto ha raggiunto terra
 	private boolean arrived = false;
 	
-	public Coin( int x, int y, float ray, double maxH ) throws SlickException
+	public Coin( float x, float y, float ray, double maxH ) throws SlickException
 		{
 			super( "coin" );
